first pass of Moai prop overlap system; added notion of overlays/underlays for rendering ad hoc layer decorations such as Box2D and Chipmunk debug lines (now that both libraries are higher level than moai-sim)

+++ src/zl-util/ZLLeanList.h
@@ -81,7 +81,7 @@ private:
 
 	ZLLeanLink < TYPE >*	mHead;
 	ZLLeanLink < TYPE >*	mTail;
-	u32						mCount;
+	size_t					mCount;
 
 public:
 
@@ -111,7 +111,7 @@ public:
 	}
 
 	//----------------------------------------------------------------//
-	u32 Count () {
+	size_t Count () {
 		return this->mCount;
 	}
 
@@ -154,7 +154,7 @@ public:
 		ZLLeanLink < TYPE >* headB = b.mHead;
 		ZLLeanLink < TYPE >* tailB = b.mTail;
 		
-		u32 count = 0;
+		size_t count = 0;
 		
 		if ( headA == headB ) {
 			count = a.mCount;
